backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic-settings==2.0.3
chromadb==0.4.15
langchain==0.0.340
langchain-community==0.0.1
google-generativeai==0.3.0
pdfplumber==0.9.0
python-docx==0.8.11
python-multipart==0.0.6
aiofiles==23.2.1 
PyPDF2==3.0.1
